+++ src/index.ts
@@ -29,6 +29,7 @@ import {
   transformConstructor,
   removeLeftoverConstructorHead,
 } from './transformations/transform-constructor';
+import { transformOptionalInitializer } from './transformations/transform-optional-initializer';
 
 interface Paths {
   root: string;
@@ -107,6 +108,7 @@ export async function transpile(
   transform.apply(transformConstructor(namespaceInclude));
   transform.apply(removeLeftoverConstructorHead);
   transform.apply(addRequiredPublicInitializer(options.publicInitializers));
+  transform.apply(transformOptionalInitializer);
   transform.apply(removeInheritanceListArguments);
   transform.apply(removeStateVarInits);
   transform.apply(removeImmutable);